Issue #11163: enforce file size on FinalLocalVariableCheckSwitchExpressions input files
