+++ src/test/resources-noncompilable/com/puppycrawl/tools/checkstyle/checks/coding/finallocalvariable/InputFinalLocalVariableCheckSwitchExpressionsB.java
@@ -1,0 +1,63 @@
+/*
+FinalLocalVariable
+validateEnhancedForLoopVariable = (default)false
+validateUnnamedVariables = (default)false
+tokens = (default)VARIABLE_DEF
+
+*/
+
+// non-compiled with eclipse: local variable may not have been initialized
+package com.puppycrawl.tools.checkstyle.checks.coding.finallocalvariable;
+
+public class InputFinalLocalVariableCheckSwitchExpressionsB {
+
+    void foo3() throws Exception {
+
+        Exception e; // violation, "Variable 'e' should be declared final"
+
+        final int a = (int) Math.random();
+        final int b = (int) Math.random();
+
+        switch (a) {
+            case 0 -> e = new Exception();
+            case 1 -> {
+                if (b == 0) {
+                    e = new Exception();
+                    break;
+                }
+
+                if (b == 1) {
+                    e = new Exception();
+                } else {
+                    e = new Exception();
+                }
+            }
+            case 2 -> {
+                if (b == 0) {
+                    return;
+                }
+
+                e = new Exception();
+            }
+            default -> e = new Exception();
+        }
+
+        throw e;
+    }
+
+    void foo4() throws Exception {
+
+        Exception e; // violation, "Variable 'e' should be declared final"
+
+        final int a = (int) Math.random();
+        final int b = (int) Math.random();
+
+        switch (a) {
+            case 0 -> e = new Exception();
+            case 1 -> System.out.println("test!");
+            default -> System.out.println("Exception!");
+
+        }
+
+    }
+}
